Moved rpc client and server to piece store
