+++ pkg/piecestore/rpc/proto/piece_store.proto
