+++ pkg/piecestore/rpc/server/api/api.go
@@ -15,10 +15,9 @@ import (
 
 	_ "github.com/mattn/go-sqlite3"
 
-	pb "storj.io/storj/pkg/rpcClientServer/protobuf"
-
 	"storj.io/storj/pkg/piecestore"
-	"storj.io/storj/pkg/rpcClientServer/server/utils"
+	pb "storj.io/storj/pkg/piecestore/rpc/proto"
+	"storj.io/storj/pkg/piecestore/rpc/server/utils"
 )
 
 // Server -- GRPC server meta data used in route calls
